#217 moved mini example modules
