+++ examples/mini-examples/fx-root-example/src/main/java/de/saxsys/jfx/mvvmfx/fx_root_example/LabeledTextFieldViewModel.java
