+++ examples/mini-examples/fx-root-example/src/main/resources/de/saxsys/jfx/mvvmfx/fx_root_example/LabeledTextField.fxml
