+++ examples/mini-examples/fx-root-example/src/test/java/de/saxsys/jfx/mvvmfx/fx_root_example/LabeledTextFieldTest.java
