+++ examples/mini-examples/helloworld-without-fxml/src/main/java/de/saxsys/jfx/mvvmfx/helloworld/HelloWorldView.java
