+++ examples/mini-examples/helloworld/src/main/java/de/saxsys/jfx/mvvmfx/helloworld/HelloWorldViewModel.java
