+++ examples/mini-examples/helloworld/src/main/resources/de/saxsys/jfx/mvvmfx/helloworld/HelloWorldView.fxml
