+++ examples/mini-examples/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<parent>
+		<artifactId>examples</artifactId>
+		<groupId>de.saxsys.mvvmfx</groupId>
+		<version>1.4.0-SNAPSHOT</version>
+	</parent>
+	<modelVersion>4.0.0</modelVersion>
+	<packaging>pom</packaging>
+	<artifactId>mini-examples</artifactId>
+
+	<name>mvvmfx - mini-examples</name>
+
+
+	<modules>
+		<module>synchronizefx-example</module>
+		<module>fx-root-example</module>
+		<module>helloworld</module>
+		<module>helloworld-without-fxml</module>
+	</modules>
+
+
+</project>
