+++ examples/mini-examples/synchronizefx-example/src/main/java/de/saxsys/mvvmfx/examples/synchronize/ClientApp.java
