+++ examples/mini-examples/synchronizefx-example/src/main/resources/de/saxsys/mvvmfx/examples/synchronize/SliderView.fxml
