--- examples/mvvmfx-fx-root-example/pom.xml
@@ -1,50 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>mvvmfx-fx-root-example</artifactId>
-	<name>Example with fx:root</name>
-
-	<parent>
-		<groupId>de.saxsys.mvvmfx</groupId>
-		<artifactId>examples</artifactId>
-		<version>1.4.0-SNAPSHOT</version>
-	</parent>
-
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
-	</properties>
-
-
-	<dependencies>
-		<dependency>
-			<groupId>de.saxsys</groupId>
-			<artifactId>mvvmfx</artifactId>
-		</dependency>
-
-		<dependency>
-			<groupId>org.assertj</groupId>
-			<artifactId>assertj-core</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>eu.lestard</groupId>
-			<artifactId>assertj-javafx</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<scope>test</scope>
-		</dependency>
-
-		<dependency>
-			<groupId>org.loadui</groupId>
-			<artifactId>testFx</artifactId>
-			<version>3.1.0</version>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-
-</project>
