--- examples/mvvmfx-synchronizefx/pom.xml
@@ -1,36 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>mvvmfx-synchronizefx</artifactId>
-	<name>SynchronizeFX example</name>
-	<parent>
-		<groupId>de.saxsys.mvvmfx</groupId>
-		<artifactId>examples</artifactId>
-		<version>1.4.0-SNAPSHOT</version>
-	</parent>
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
-	</properties>
-
-
-	<dependencies>
-		<dependency>
-			<groupId>de.saxsys</groupId>
-			<artifactId>mvvmfx</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>de.saxsys.synchronizefx</groupId>
-			<artifactId>synchronizefx</artifactId>
-			<version>0.4.0</version>
-			<exclusions>
-				<exclusion>
-					<groupId>com.oracle</groupId>
-					<artifactId>javafx</artifactId>
-				</exclusion>
-			</exclusions>
-		</dependency>
-	</dependencies>
-
-</project>
