+++ examples/pom.xml
@@ -17,18 +17,20 @@
 		<module>mvvmfx-complex-example</module>
 		<module>mvvmfx-guice-starter</module>
 		<module>mvvmfx-cdi-starter</module>
-		<module>mvvmfx-fx-root-example</module>
-		<module>mvvmfx-helloworld</module>
-		<module>mvvmfx-helloworld-without-fxml</module>
-		<module>mvvmfx-synchronizefx</module>
 		<module>contacts-example</module>
 		<module>books-example</module>
 		<module>todomvc-example</module>
+		<module>mini-examples</module>
 	</modules>
 
 
 	<dependencyManagement>
 		<dependencies>
+			<dependency>
+				<groupId>de.saxsys.mvvmfx</groupId>
+				<artifactId>mini-examples</artifactId>
+				<version>${project.version}</version>
+			</dependency>
 			<dependency>
 				<groupId>de.saxsys.mvvmfx</groupId>
 				<artifactId>mvvmfx-complex</artifactId>
@@ -46,22 +48,22 @@
 			</dependency>
 			<dependency>
 				<groupId>de.saxsys.mvvmfx</groupId>
-				<artifactId>mvvmfx-fx-root-example</artifactId>
+				<artifactId>fx-root-example</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>de.saxsys.mvvmfx</groupId>
-				<artifactId>mvvmfx-helloworld</artifactId>
+				<artifactId>helloworld</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>de.saxsys.mvvmfx</groupId>
-				<artifactId>mvvmfx-helloworld-without-fxml</artifactId>
+				<artifactId>helloworld-without-fxml</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
 				<groupId>de.saxsys.mvvmfx</groupId>
-				<artifactId>mvvmfx-synchronizefx</artifactId>
+				<artifactId>synchronizefx-example</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
